requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# MongoDB
motor==3.3.1
pymongo==4.6.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic[email]==2.5.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
bson==0.5.10
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code Quality and Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Type Stubs
types-python-jose==3.3.4
types-passlib==1.7.7
types-pytz==2023.3.1.1
types-python-dateutil==2.8.19.14

# Email
aiosmtplib==2.0.2